remove header for PCLMetadataWriter plugin and do ESconsumes migration
